Move the SlintContext to a different module
