+++ internal/core/input.rs
@@ -499,7 +499,7 @@ impl ClickState {
 
                 if let Some(click_count_time_stamp) = self.click_count_time_stamp.get() {
                     if instant_now - click_count_time_stamp
-                        < crate::GLOBAL_CONTEXT
+                        < crate::context::GLOBAL_CONTEXT
                             .with(|p| p.get().map(|p| p.platform.click_interval()))
                             .unwrap_or_default()
                         && button == self.click_button.get()
